Added micropython example code
Tested only with Micropython ESP32 on the WifiKit32 controller

Tested against both Mac and iPhone.

Interesting, the "name" of the BLE connection is correct on the Mac, but on the iPhone it is always showing "ESP32" (some hard-coded value in the micropython code).  More investigation needed here, but finding the service by the ID works as expected.

Next is to have the OLED on the ESP32 to display the value sent by the Hello-Bluetooth app.